refactoring: split parse_typemap_macro to typemap_macro and parse
