+++ macroslib/src/typemap/parse.rs
@@ -20,8 +20,8 @@ use crate::{
             normalize_ty_lifetimes, parse_ty_with_given_span, DisplayToTokens, GenericTypeConv,
             TypeName,
         },
-        parse_typemap_macro::TypeMapConvRuleInfo,
         ty::{ForeignTypesStorage, RustTypeS},
+        typemap_macro::TypeMapConvRuleInfo,
         validate_code_template, TypeConvEdge, TypeMap, TypesConvGraph,
     },
     FOREIGN_TYPEMAP,
